Adapt to new dir structure
