+++ experimental/PlaneCurve/Variety.jl
